+++ jsp/resign_action.jsp
@@ -8,14 +8,21 @@
     request.setCharacterEncoding("utf-8");
     Object id_value_ob = session.getAttribute("id_value");
     String id_value = String.valueOf(id_value_ob);
-
-    Class.forName("com.mysql.jdbc.Driver");
-    Connection connect = DriverManager.getConnection("jdbc:mysql://localhost/week09","gongsil","1005");
-
-    String sql = "DELETE FROM user WHERE id= ?";
-    PreparedStatement query = connect.prepareStatement(sql);
-    query.setString(1, id_value);
-    query.executeUpdate();
+    if(id_value_ob == null){
+        response.sendRedirect("/week09/jsp/log_in.jsp");
+        return;
+    }
+    try{
+        Class.forName("com.mysql.jdbc.Driver");
+        Connection connect = DriverManager.getConnection("jdbc:mysql://localhost/week09","gongsil","1005");
+    
+        String sql = "DELETE FROM user WHERE id= ?";
+        PreparedStatement query = connect.prepareStatement(sql);
+        query.setString(1, id_value);
+        query.executeUpdate();
+    }catch(Exception e){
+        e.printStackTrace();
+    }
 %>
 
 <head>
